api/pyproject.toml
[project]
name = "hisoka-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.15.2",
    "fastapi[standard]>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pwdlib[argon2]>=0.2.1",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.1",
    "pytest-azurepipelines>=1.0.5",
    "sqlalchemy>=2.0.40",
    "tzdata>=2025.2",
]

[tool.ruff]
line-length = 128
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev app/main.py'
pre_test = 'task format'
test = 'pytest -s -x --junitxml=junit/test-results.xml --cov=app --cov-report=xml -vv '
post_test = 'coverage html'

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true

[dependency-groups]
dev = [
    "commitizen>=4.4.1",
    "factory-boy>=3.3.3",
    "freezegun>=1.5.1",
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.0",
    "ruff>=0.11.3",
    "taskipy>=1.14.1",
    "testcontainers>=4.10.0",
]
